+++ resolwe/flow/views/descriptor.py
@@ -5,7 +5,7 @@ from rest_framework import mixins, viewsets
 
 from resolwe.flow.models import DescriptorSchema
 from resolwe.flow.serializers import DescriptorSchemaSerializer
-from resolwe.permissions.loader import permissions_cls
+from resolwe.permissions.loader import get_permissions_class
 from resolwe.permissions.mixins import ResolwePermissionsMixin
 
 
@@ -17,7 +17,7 @@ class DescriptorSchemaViewSet(mixins.RetrieveModelMixin,
 
     queryset = DescriptorSchema.objects.all().prefetch_related('contributor')
     serializer_class = DescriptorSchemaSerializer
-    permission_classes = (permissions_cls,)
+    permission_classes = (get_permissions_class(),)
     filter_fields = ('contributor', 'name', 'description', 'created', 'modified', 'slug')
     ordering_fields = ('id', 'created', 'modified', 'name', 'version')
     ordering = ('id',)